+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for Active/Passive deployment, [PR-1028](https://github.com/reductstore/reductstore/pull/1028)
 - Implement concurrent read-only access to data, [PR-1034](https://github.com/reductstore/reductstore/pull/1034)
+- Cache directory tree
 
 ### Internal