+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 1.17.7 - 2025-12-12
+
+### Fixed
+
+- Fix deadlock in storage engine, using RwLock with timeout, [PR-1059](https://github.com/reductstore/reductstore/pull/1059)
+
 ## 1.17.6 - 2025-12-03
 
 ### Fixed